Split up implementation traits for abstract spetroscopy and abstract individual eq checking
