+++ js-client/src/main/scala-2.12/io/equiv/eqfiddle/tool/control/Structure.scala
@@ -303,7 +303,7 @@ object Structure {
         }
         AlgorithmLogging.uriEncoder = scala.scalajs.js.URIUtils.encodeURI _
 
-        val result = algo.compute(List((n1, n2)), SpectroscopyInterface.SpectroscopyConfig(computeFormulas = true))
+        val result = algo.decideAll(List((n1, n2)), SpectroscopyInterface.SpectroscopyConfig(computeFormulas = true))
         AlgorithmLogging.debugLog("Spectroscopy took: " + (Date.now - begin) + "ms.", logLevel = 7)
 
         val gameString = result.meta.get("game") match {
@@ -427,7 +427,7 @@ object Structure {
         n2j <- (n1i + 1) until states.length
       } yield (states(n1i), states(n2j))
 
-      val result = algo.compute(comparedPairs, SpectroscopyInterface.SpectroscopyConfig(computeFormulas = false, energyCap = 3))
+      val result = algo.decideAll(comparedPairs, SpectroscopyInterface.SpectroscopyConfig(computeFormulas = false, energyCap = 3))
       AlgorithmLogging.debugLog("Minimization Spectroscopy took: " + (Date.now - begin) + "ms.", logLevel = 7)
 
       val distRel = result.toDistancesRelation()
@@ -470,7 +470,7 @@ object Structure {
           n2 <- structure.structure.nodes
         } yield (node, n2)
 
-        val result = algo.compute(comparedPairs, SpectroscopyInterface.SpectroscopyConfig(computeFormulas = false, energyCap = 3))
+        val result = algo.decideAll(comparedPairs, SpectroscopyInterface.SpectroscopyConfig(computeFormulas = false, energyCap = 3))
         AlgorithmLogging.debugLog("Characterization Spectroscopy took: " + (Date.now - begin) + "ms.", logLevel = 7)
 
         for {
