+++ shared/src/main/scala-2.12/io/equiv/eqfiddle/spectroscopy/WeakSpectroscopy.scala
@@ -16,10 +16,11 @@ import io.equiv.eqfiddle.game.MaterializedEnergyGame
 
 class WeakSpectroscopy[S, A, L] (
     override val ts: WeakTransitionSystem[S, A, L])
-  extends SpectroscopyInterface[S, A, L, HennessyMilnerLogic.Formula[A]] {
+  extends SpectroscopyInterface[S, A, L, HennessyMilnerLogic.Formula[A]]
+  with SpectroscopyFramework[S, A, L, HennessyMilnerLogic.Formula[A]] 
+  with SpectroscopyEquivalenceChecking[S, A, L, HennessyMilnerLogic.Formula[A]] {
 
   import WeakSpectroscopyGame._
-  import MaterializedEnergyGame._
 
   type Notion = ObservationNotionWeak
   type GamePosition = WeakSpectroscopyGamePosition[S, A]
