+++ shared/src/test/scala-2.12/io/equiv/eqfiddle/CSSSampleTests.scala
@@ -54,7 +54,7 @@ trait CSSSampleTests[OC <: ObservationNotion, CF <: HennessyMilnerLogic.Formula[
           val preordsStr = preords.map(_.name)
           val notPreordsStr = notPreords.map(_.name).intersect(algo.spectrum.notionNames)
 
-          val result = algo.compute(List((n1, n2)), spectroscopyConfig)
+          val result = algo.decideAll(List((n1, n2)), spectroscopyConfig)
 
           val foundDistinctions = result.foundDistinctions(n1, n2).map(
             d => d.name match { case "2bisimulation" => "bisimulation"; case "2trace" => "trace"; case n => n }
